requirements.txt
torchvision
torch
scipy
wandb
lmdb
numpy
simple-parsing
pandas
pillow
einops
matplotlib
tqdm